Docs: updated demo (#19)
