+++ pytest.ini
@@ -1,0 +1,7 @@
+# Setup for pytest
+[pytest]
+markers =
+    mat_ops: mark a test as utilizing matrix operations.
+; addopts = --durations=10
+addopts = --doctest-modules
+doctest_optionflags = NORMALIZE_WHITESPACE IGNORE_EXCEPTION_DETAIL
